refactored sortQueryParameters into RasQueryParameters, changing all references to it in the process.
