+++ galasa-parent/dev.galasa.framework.api.ras/src/main/java/dev/galasa/framework/api/ras/internal/routes/ResultNamesRoute.java
@@ -16,8 +16,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
-import dev.galasa.framework.api.ras.internal.verycommon.*;
-import dev.galasa.framework.api.ras.internal.common.SortQueryParameterChecker;
+import dev.galasa.framework.api.ras.internal.common.RasQueryParameters;
 import dev.galasa.framework.api.ras.internal.verycommon.InternalServletException;
 import dev.galasa.framework.api.ras.internal.verycommon.QueryParameters;
 import dev.galasa.framework.api.ras.internal.verycommon.ResponseBuilder;
@@ -39,7 +38,6 @@ public class ResultNamesRoute extends RunsRoute {
 	}
 
 	final static Gson gson = GalasaGsonBuilder.build();
-    private SortQueryParameterChecker sortQueryParameterChecker = new SortQueryParameterChecker();
 
     @Override
     public HttpServletResponse handleRequest(String pathInfo, QueryParameters queryParams, HttpServletResponse response) throws ServletException, IOException, FrameworkException {
@@ -49,10 +47,11 @@ public class ResultNamesRoute extends RunsRoute {
 
     public String retrieveResults (QueryParameters queryParams) throws ServletException, InternalServletException{
         List<String> resultsList = getResultNames();
-
+		RasQueryParameters rQueryParameters = new RasQueryParameters(queryParams);
+	
 		try {
             if (queryParams.getSingleString("sort", null) !=null ){
-			    if (!sortQueryParameterChecker.isAscending(queryParams, "resultnames")) {
+			    if (!rQueryParameters.isAscending("resultnames")) {
 				    Collections.reverse(resultsList);
                 }
 			}
